+++ src/components/SearchModal/CurrencyList.tsx
@@ -17,7 +17,7 @@ import { MouseoverTooltip } from '../Tooltip'
 import { MenuItem } from './styleds'
 import Loader from '../Loader'
 import { isTokenOnList } from 'utils'
-import ImportRow from 'components/SearchModal/CurrencyList/ImportRow'
+import ImportRow from 'components/SearchModal/ImportRow'
 import { LightGreyCard } from 'components/Card'
 import TokenListLogo from '../../assets/svg/tokenlist.svg'
 import QuestionHelper from 'components/QuestionHelper'
