+++ src/custom/components/SearchModal/CurrencyList/CurrencyListMod.tsx
@@ -17,7 +17,7 @@ import { MouseoverTooltip } from 'components/Tooltip'
 import { MenuItem } from 'components/SearchModal/styleds'
 import Loader from 'components/Loader'
 import { isTokenOnList } from 'utils'
-import ImportRow from 'components/SearchModal/CurrencyList/ImportRow'
+import ImportRow from 'components/SearchModal/ImportRow'
 import { LightGreyCard } from 'components/Card'
 import TokenListLogo from 'assets/svg/tokenlist.svg'
 import QuestionHelper from 'components/QuestionHelper'
